added css, js. and changed things in html
Added css to change background color, h1 color and changed the placement of it. Also added css for the different checkmark. Added js for the checkmark when clicked. Changed the placement of the js script in the html file to the top as well as added a div on top of the checkmark.